+++ src/renderer/component/common/form-components/form-field.jsx
@@ -23,7 +23,6 @@ type Props = {
   children?: React.Node,
   stretch?: boolean,
   affixClass?: string, // class applied to prefix/postfix label
-  useToggle?: boolean,
   firstInList?: boolean, // at the top of a list, no padding top
 };
 
@@ -41,7 +40,6 @@ export class FormField extends React.PureComponent<Props> {
       children,
       stretch,
       affixClass,
-      useToggle,
       firstInList,
       ...inputProps
     } = this.props;
@@ -79,7 +77,7 @@ export class FormField extends React.PureComponent<Props> {
         );
       } else if (type === 'textarea') {
         input = <textarea type={type} id={name} {...inputProps} />;
-      } else if (type === 'checkbox' && useToggle) {
+      } else if (type === 'checkbox') {
         input = <Toggle id={name} {...inputProps} />;
       } else {
         input = <input type={type} id={name} {...inputProps} />;