+++ src/renderer/constants/settings.js
@@ -14,3 +14,4 @@ export const THEMES = 'themes';
 export const AUTOMATIC_DARK_MODE_ENABLED = 'automaticDarkModeEnabled';
 export const AUTOPLAY = 'autoplay';
 export const RESULT_COUNT = 'resultCount';
+export const OS_NOTIFICATIONS_ENABLED = 'osNotificationsEnabled';