+++ src/renderer/redux/reducers/settings.js
@@ -26,6 +26,9 @@ const defaultState = {
     automaticDarkModeEnabled: getLocalStorageSetting(SETTINGS.AUTOMATIC_DARK_MODE_ENABLED, false),
     autoplay: getLocalStorageSetting(SETTINGS.AUTOPLAY, false),
     resultCount: Number(getLocalStorageSetting(SETTINGS.RESULT_COUNT, 50)),
+    osNotificationsEnabled: Boolean(
+      getLocalStorageSetting(SETTINGS.OS_NOTIFICATIONS_ENABLED, true)
+    ),
   },
   isNight: false,
   languages: {},